feat : user정보를 전역상태로 사용할 수 있도록 변경
