+++ front/src/pages/Main/index.tsx
@@ -1,8 +1,5 @@
 import React, { useEffect, useState, useRef } from 'react';
 import styled from 'styled-components';
-import { useSetRecoilState } from 'recoil';
-import { loginState } from '../../recoil/atom';
-import { getUser } from '../../utils/api/auth';
 import Header from '../../components/Header';
 import Keywords from '../../components/Keywords';
 import WorldCupList from '../../components/WorldcupList';
@@ -12,7 +9,6 @@ enum filtering {
   search,
 }
 function Main(): JSX.Element {
-  const setIsLoggedIn = useSetRecoilState(loginState);
   const [inputWord, setInputWord] = useState('');
   const [searchWord, setSearchWord] = useState('');
   const [selectedTag, setSelectedTag] = useState('');
@@ -33,12 +29,6 @@ function Main(): JSX.Element {
     setSelectedTag(keyword);
     filterStandard.current = 0;
   };
-  const getUserInfo = async () => {
-    const user = await getUser();
-    if (Object.keys(user).length !== 0) {
-      setIsLoggedIn(true);
-    }
-  };
   useEffect(() => {
     const script = document.createElement('script');
     script.src = 'https://developers.kakao.com/sdk/js/kakao.js';
@@ -48,9 +38,7 @@ function Main(): JSX.Element {
       document.body.removeChild(script);
     };
   }, []);
-  useEffect(() => {
-    getUserInfo();
-  }, []);
+
   return (
     <Wrapper>
       <Header type="searchHeader" onSubmit={onSubmit} onSearchWordChange={onSearchWordChange} searchWord={inputWord} />
